changed float to double for variables in main function

+++ lab1/presekNaOtsecki.c
@@ -60,14 +60,14 @@ int se_secat(otsecka o1, otsecka o2) {
 }
 
 int main() {
-    float x1, y1, x2, y2;
-    scanf("%f %f %f %f", &x1, &y1, &x2, &y2);
+    double x1, y1, x2, y2;
+    scanf("%lf %lf %lf %lf", &x1, &y1, &x2, &y2);
 
     tocka t1 = {x1, y1};
     tocka t2 = {x2, y2};
     otsecka o1 = {t1, t2};
 
-    scanf("%f %f %f %f", &x1, &y1, &x2, &y2);
+    scanf("%lf %lf %lf %lf", &x1, &y1, &x2, &y2);
 
     tocka t3 = {x1, y1};
     tocka t4 = {x2, y2};
